+++ packages/lumx-react/src/index.ts
@@ -10,6 +10,7 @@ export * from './components/autocomplete';
 export * from './components/avatar';
 export * from './components/badge';
 export * from './components/button';
+export * from './components/card';
 export * from './components/checkbox';
 export * from './components/chip';
 export * from './components/combobox';